# internal/api/http/openapi.yaml
openapi: 3.1.0
info:
  title: Plowered API
  version: "1.0.0"
  description: |
    Plowered is a self-hosted, AI-native data context platform: catalog,
    orchestration, quality, governance, lineage, ETL/ELT, MCP, all in one.

    All endpoints sit behind the same auth chain — session cookie (web UI)
    or Bearer JWT (integrations). Tenant isolation is enforced server-side
    from the authenticated principal; you never pass tenant_id in payloads
    or query strings.

    Identifiers throughout are UUIDs unless documented otherwise.
  contact:
    name: Plowered
    url: https://github.com/Satyaamm/plowered
servers:
  - url: http://localhost:8080
    description: Local dev
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /v1/assets:
    get:
      tags: [catalog]
      summary: List assets (paginated)
      parameters:
        - in: query
          name: type
          schema: { type: string, enum: [database, schema, table, view, column, dashboard, report, dbt_model, dag, ml_model, glossary_term] }
        - in: query
          name: page_size
          schema: { type: integer, default: 50, maximum: 500 }
        - in: query
          name: page_token
          schema: { type: string }
      responses:
        "200":
          description: Asset page
          content:
            application/json:
              schema:
                type: object
                properties:
                  assets:
                    type: array
                    items: { $ref: "#/components/schemas/Asset" }
                  next_page_token: { type: string }
    post:
      tags: [catalog]
      summary: Upsert an asset by qualified_name
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/Asset" } } }
      responses:
        "201": { description: Asset created or replaced }
  /v1/assets/{id}:
    get:
      tags: [catalog]
      summary: Get an asset by UUID
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: Asset, content: { application/json: { schema: { $ref: "#/components/schemas/Asset" } } } }
        "404": { description: Not found }
  /v1/assets/{id}/lineage:
    get:
      tags: [catalog]
      summary: Get neighbouring lineage edges
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: direction, in: query, schema: { type: string, enum: [upstream, downstream, both], default: both } }
        - { name: depth, in: query, schema: { type: integer, default: 2, maximum: 5 } }
        - { name: kind, in: query, schema: { type: string, enum: [defines, lineage] } }
      responses:
        "200": { description: Neighbourhood }
...
  /v1/assets:byQualifiedName:
    get:
      tags: [catalog]
      summary: Get an asset by qualified name
      parameters:
        - { name: qn, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: Asset, content: { application/json: { schema: { $ref: "#/components/schemas/Asset" } } } }
        "404": { description: Not found }
  /v1/assets:search:
    post:
      tags: [catalog]
      summary: Keyword search across qualified names
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query: { type: string }
                limit: { type: integer, default: 20, maximum: 200 }
                type: { type: string }
      responses:
        "200": { description: Hits }

  # ---------- search (semantic) ----------
  /v1/search:semantic:
    post:
      tags: [ai]
      summary: Vector-similarity search filtered by policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query: { type: string }
                k: { type: integer, default: 20, maximum: 100 }
      responses:
        "200":
          description: Ranked hits
          content:
            application/json:
              schema:
                type: object
                properties:
                  query: { type: string }
                  hits:
                    type: array
                    items: { $ref: "#/components/schemas/SemanticHit" }
  /v1/search:reindex:
    post:
      tags: [ai]
      summary: Re-embed every asset in the tenant
      description: |
        When the platform is wired with a Jobs repo + Asynq enqueuer
        (production), returns 202 + a job_id and the actual reindex runs
        on a worker. Poll /v1/jobs/{id} for progress + final result. In
        embedded/dev mode (no jobs wiring) the endpoint runs synchronously
        and returns the legacy 200 + {reindexed, model} body.
      responses:
        "200":
          description: Reindex summary (legacy synchronous path)
          content:
            application/json:
              schema:
                type: object
                properties:
                  reindexed: { type: integer }
                  model: { type: string }
        "202":
          description: Queued — poll /v1/jobs/{id}
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id: { type: string, format: uuid }
                  status: { type: string }
                  model:  { type: string }

  # ---------- connections ----------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: plowered_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Asset:
      type: object
      properties:
        id:             { type: string, format: uuid }
        tenant_id:      { type: string, format: uuid }
        qualified_name: { type: string }
        type:           { type: string }
        name:           { type: string }
        description:    { type: string }
        description_ai: { type: string }
        trust:          { type: string, enum: [unverified, draft, reviewed, certified, deprecated] }
        tags:           { type: array, items: { type: string } }
        owners:         { type: array, items: { type: string } }
        properties:     { type: object, additionalProperties: true }
        created_at:     { type: string, format: date-time }
        updated_at:     { type: string, format: date-time }
...
    SemanticHit:
      type: object
      properties:
        asset_id:       { type: string, format: uuid }
        qualified_name: { type: string }
        type:           { type: string }
        description:    { type: string }
        tags:           { type: array, items: { type: string } }
        score:          { type: number, format: float }
...